CON-359: Move finality-specific code to new package.
