+++ casper/src/main/scala/io/casperlabs/casper/api/BlockAPI.scala
@@ -12,6 +12,7 @@ import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
 import io.casperlabs.casper.consensus._
 import io.casperlabs.casper.consensus.info._
 import io.casperlabs.casper.deploybuffer.DeployBuffer
+import io.casperlabs.casper.finality.FinalityDetector
 import io.casperlabs.casper.protocol.{
   BlockInfoWithoutTuplespace,
   BlockQuery,
