+++ casper/src/test/scala/io/casperlabs/casper/api/BlocksResponseAPITest.scala
@@ -7,6 +7,7 @@ import com.google.protobuf.ByteString
 import io.casperlabs.blockstorage.{BlockStorage, IndexedDagStorage}
 import io.casperlabs.casper.Estimator.BlockHash
 import io.casperlabs.casper.consensus._
+import io.casperlabs.casper.finality.FinalityDetectorBySingleSweepImpl
 import io.casperlabs.casper.{genesis, _}
 import io.casperlabs.casper.helper._
 import io.casperlabs.casper.helper.BlockGenerator._
