+++ casper/src/test/scala/io/casperlabs/casper/finality/FinalityDetectorBySingleSweepTest.scala
@@ -1,12 +1,12 @@
-package io.casperlabs.casper
+package io.casperlabs.casper.finality
 
 import com.github.ghik.silencer.silent
 import com.google.protobuf.ByteString
 import io.casperlabs.casper.consensus.Bond
-import io.casperlabs.casper.helper.{BlockGenerator, DagStorageFixture}
+import io.casperlabs.casper.finality.FinalityDetector.Committee
 import io.casperlabs.casper.helper.BlockGenerator._
 import io.casperlabs.casper.helper.BlockUtil.generateValidator
-import io.casperlabs.casper.FinalityDetector.Committee
+import io.casperlabs.casper.helper.{BlockGenerator, DagStorageFixture}
 import io.casperlabs.casper.scalatestcontrib._
 import io.casperlabs.p2p.EffectsTestInstances.LogStub
 import monix.eval.Task
