+++ casper/src/test/scala/io/casperlabs/casper/helper/HashSetCasperTestNode.scala
@@ -25,7 +25,8 @@ import io.casperlabs.crypto.Keys
 import io.casperlabs.crypto.Keys.{PrivateKey, PublicKey}
 import io.casperlabs.crypto.signatures.SignatureAlgorithm.Ed25519
 import io.casperlabs.ipc
-import io.casperlabs.casper.consensus.state.{Unit => _, BigInt => _, _}
+import io.casperlabs.casper.consensus.state.{BigInt => _, Unit => _, _}
+import io.casperlabs.casper.finality.FinalityDetector
 import io.casperlabs.casper.validation.Validation
 import io.casperlabs.ipc.DeployResult.Value.ExecutionResult
 import io.casperlabs.ipc.TransformEntry
