+++ node/src/main/scala/io/casperlabs/node/api/GrpcDeployService.scala
@@ -8,8 +8,8 @@ import com.google.protobuf.ByteString
 import com.google.protobuf.empty.Empty
 import io.casperlabs.blockstorage.BlockStorage
 import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
-import io.casperlabs.casper.FinalityDetector
 import io.casperlabs.casper.api.BlockAPI
+import io.casperlabs.casper.finality.FinalityDetector
 import io.casperlabs.casper.protocol._
 import io.casperlabs.catscontrib.TaskContrib._
 import io.casperlabs.comm.ServiceError.Unimplemented
