+++ node/src/main/scala/io/casperlabs/node/api/Servers.scala
@@ -6,9 +6,9 @@ import cats.effect.{Effect => _, _}
 import cats.implicits._
 import io.casperlabs.blockstorage.BlockStorage
 import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
-import io.casperlabs.casper.FinalityDetector
 import io.casperlabs.casper.deploybuffer.DeployBuffer
 import io.casperlabs.casper.consensus.Block
+import io.casperlabs.casper.finality.FinalityDetector
 import io.casperlabs.casper.protocol.CasperMessageGrpcMonix
 import io.casperlabs.casper.validation.Validation
 import io.casperlabs.comm.discovery.{NodeDiscovery, NodeIdentifier}
