+++ node/src/main/scala/io/casperlabs/node/api/graphql/schema/GraphQLSchemaBuilder.scala
@@ -3,11 +3,11 @@ package io.casperlabs.node.api.graphql.schema
 import cats.implicits._
 import io.casperlabs.blockstorage.BlockStorage
 import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
-import io.casperlabs.casper.FinalityDetector
 import io.casperlabs.casper.api.BlockAPI
 import io.casperlabs.catscontrib.MonadThrowable
 import io.casperlabs.casper.consensus.state
 import io.casperlabs.casper.deploybuffer.DeployBuffer
+import io.casperlabs.casper.finality.FinalityDetector
 import io.casperlabs.models.SmartContractEngineError
 import io.casperlabs.node.api.Utils
 import io.casperlabs.node.api.graphql.RunToFuture.ops._
